Allow selecting documents using tags
Tags can be viewed as keywords an can be used to aid on selecting
documents which might be related.

This change adds a list of all tags available in ansible-freeipa
documentation after the 'Quick Links' section, and a new page 'tags'
containing a list of tags with links to all documents that contain the
tag.

Tag lists are sorted.